requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
defusedxml==0.7.1
elasticsearch==7.13.4
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
limits==1.5.1
MarkupSafe==2.0.1
pbr==5.6.0
pylogbeat==2.0.0
python-logstash-async==2.3.0
redis==3.5.3
requests==2.26.0
servicenow==2.1.0
six==1.16.0
SOAPpy==0.12.22
urllib3==1.26.6
Werkzeug==2.0.1
wstools==0.4.8
